requirements.txt
appnope==0.1.2
autopep8==1.5.6
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
greenlet==1.0.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.3.4
nltk==3.5
numpy==1.20.0
pandas==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
praw==7.2.0
prawcore==2.0.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyzmq==22.0.2
regex==2021.3.17
requests==2.25.1
schedule==1.1.0
six==1.15.0
SQLAlchemy==1.4.11
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
update-checker==0.18.0
urllib3==1.26.3
wcwidth==0.2.5
websocket-client==0.57.0
wordcloud==1.8.1
